tianyan-dashou/agents/data-agent/requirements.txt
# Data Agent 依赖文件

# 核心依赖
requests==2.31.0
schedule==1.2.1
prefect==2.14.13

# 遥感处理
gdal==3.7.3
rasterio==1.3.8
numpy==1.24.4
scipy==1.10.1

# 空间数据
shapely==2.0.1
pyproj==3.6.1

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# 配置
yaml==0.2.5
python-dotenv==1.0.0

# 工具
tqdm==4.66.1
loguru==0.7.2